Increased console logging, temp max now 84C, and suppress HTTP post to emoncms when json is empty.
All changes marked with JCC.  The two most useful things to review are use of date processing to timestamp some log messages and a pre-ECMA5 routine scraped from web for checking if a javascript object is empty.
The empty json arises when a reading is outside set min/max limits; a more direct fix would check where the json is constructed but that's harder because multiple readings get packed in one json structure.